centering refactoring
